# Optimize MVCCGet path.