Provide fast-paths for retrieving unmarshalling proto values that do not
copy the data into a Go slice. Avoid unnecessary allocations when
encoding mvcc keys.

benchmark                             old ns/op     new ns/op     delta
BenchmarkMVCCScan1Version1Row         11322         10721         -5.31%
BenchmarkMVCCScan1Version10Rows       36566         27190         -25.64%
BenchmarkMVCCScan1Version100Rows      276241        217066        -21.42%
BenchmarkMVCCScan1Version1000Rows     2205243       1834826       -16.80%
BenchmarkMVCCGet1Version              3739          3153          -15.67%

benchmark                             old MB/s     new MB/s     speedup
BenchmarkMVCCScan1Version1Row         90.44        95.51        1.06x
BenchmarkMVCCScan1Version10Rows       280.04       376.60       1.34x
BenchmarkMVCCScan1Version100Rows      370.69       471.74       1.27x
BenchmarkMVCCScan1Version1000Rows     464.35       558.09       1.20x
BenchmarkMVCCGet1Version              273.83       324.77       1.19x