@petermattis 2015-03-24
Not quite ready for another look. I have to track down the stuff about a 2GB data limit and what the appropriate constant is to use there.